Move RemappingPicker
