+++ frontend/src/metabase/metadata/components/index.ts
@@ -7,7 +7,6 @@ export * from "./FieldValuesTypePicker";
 export * from "./FieldVisibilityPicker";
 export * from "./FkTargetPicker";
 export * from "./NameDescriptionInput";
-export * from "./RemappingPicker";
 export * from "./RescanFieldButton";
 export * from "./RescanTableFieldsButton";
 export * from "./SemanticTypeAndTargetPicker";
