+++ frontend/src/metabase/metadata/pages/DataModel/components/FieldSection/BehaviorSection/index.ts
@@ -1,0 +1,1 @@
+export * from "./BehaviorSection";
